requirements.txt
django
djangorestframework
pre-commit
ipython
